Implemented OrderController and OrderService; Refactored all double values to BigDecimal for accurate financial calculations; Implemented global exception handling for duplicate emails and invalid addresses

+++ pipes-bakery-backend/src/main/java/com/hyd/pipes_bakery_backend/dto/orderItem/OrderItemRequestDTO.java
@@ -1,5 +1,7 @@
 package com.hyd.pipes_bakery_backend.dto.orderItem;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 
@@ -14,7 +16,7 @@ public class OrderItemRequestDTO {
 
     @NotBlank(message = "Unit price at purchase is required")
     @Min(value = 0, message = "Unit price at purchase must be at least 0")
-    private double unitPriceAtPurchase;
+    private BigDecimal unitPriceAtPurchase;
 
 
     public long getProductId() {
@@ -23,7 +25,7 @@ public class OrderItemRequestDTO {
     public int getQuantity() {
         return quantity;
     }
-    public double getUnitPriceAtPurchase() {
+    public BigDecimal getUnitPriceAtPurchase() {
         return unitPriceAtPurchase;
     }
     public void setProductId(long productId) {
@@ -32,7 +34,7 @@ public class OrderItemRequestDTO {
     public void setQuantity(int quantity) {
         this.quantity = quantity;
     }
-    public void setUnitPriceAtPurchase(double unitPriceAtPurchase) {
+    public void setUnitPriceAtPurchase(BigDecimal unitPriceAtPurchase) {
         this.unitPriceAtPurchase = unitPriceAtPurchase;
     }
 }
